Write uint16 files where possible (need to extend this to bioformats) - significant speed increase loading data

+++ FLIMfitFrontEnd/Classes/@flim_data_series/flim_data_series.m
@@ -103,7 +103,7 @@ classdef flim_data_series < handle & h5_serializer
         n_chan = 1;
         polarisation_resolved = false;
         data_size;
-        data_type = 'uint16';
+        data_type = 'single';
         use;
         
         use_smoothing = false;
